+++ src/Schema/Directives/LazyLoadDirective.php
@@ -3,14 +3,16 @@
 namespace Nuwave\Lighthouse\Schema\Directives;
 
 use Closure;
-use GraphQL\Deferred;
-use GraphQL\Type\Definition\ResolveInfo;
+use GraphQL\Language\AST\FieldDefinitionNode;
+use GraphQL\Language\AST\ObjectTypeDefinitionNode;
 use Illuminate\Database\Eloquent\Collection;
+use Nuwave\Lighthouse\Exceptions\DefinitionException;
+use Nuwave\Lighthouse\Schema\AST\DocumentAST;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
+use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
-class LazyLoadDirective extends BaseDirective implements FieldMiddleware
+class LazyLoadDirective extends BaseDirective implements FieldMiddleware, FieldManipulator
 {
     public static function definition(): string
     {
@@ -30,26 +32,24 @@ GRAPHQL;
 
     public function handleField(FieldValue $fieldValue, Closure $next): FieldValue
     {
-        $relations = $this->directiveArgValue('relations', []);
-        $resolver = $fieldValue->getResolver();
+        $relations = $this->directiveArgValue('relations');
 
-        return $next(
-            $fieldValue->setResolver(
-                function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($resolver, $relations) {
-                    /** @var \GraphQL\Deferred|\Illuminate\Database\Eloquent\Model $result */
-                    $result = $resolver($root, $args, $context, $resolveInfo);
+        $fieldValue->resultHandler(static function (Collection $items) use ($relations): Collection {
+            $items->load($relations);
 
-                    $result instanceof Deferred
-                        ? $result->then(function (Collection &$items) use ($relations): Collection {
-                            $items->load($relations);
+            return $items;
+        });
 
-                            return $items;
-                        })
-                        : $result->load($relations);
+        return $next($fieldValue);
+    }
 
-                    return $result;
-                }
-            )
-        );
+    public function manipulateFieldDefinition(DocumentAST &$documentAST, FieldDefinitionNode &$fieldDefinition, ObjectTypeDefinitionNode &$parentType)
+    {
+        $relations = $this->directiveArgValue('relations');
+        if (! is_array($relations) || count($relations) === 0) {
+            throw new DefinitionException(
+                "Must specify non-empty list of relations in `@{$this->name()}` directive on `{$parentType->name->value}.{$fieldDefinition->name->value}`."
+            );
+        }
     }
 }