+++ src/Schema/Directives/WithCountDirective.php
@@ -2,11 +2,18 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
+use GraphQL\Language\AST\FieldDefinitionNode;
+use GraphQL\Language\AST\ObjectTypeDefinitionNode;
+use GraphQL\Type\Definition\ResolveInfo;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
-use Nuwave\Lighthouse\Execution\DataLoader\RelationCountBatchLoader;
+use Nuwave\Lighthouse\Execution\DataLoader\RelationCountLoader;
+use Nuwave\Lighthouse\Execution\DataLoader\RelationLoader;
+use Nuwave\Lighthouse\Schema\AST\DocumentAST;
+use Nuwave\Lighthouse\Schema\RootType;
+use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 
-class WithCountDirective extends WithRelationDirective implements FieldMiddleware
+class WithCountDirective extends WithRelationDirective implements FieldMiddleware, FieldManipulator
 {
     public static function definition(): string
     {
@@ -32,18 +39,34 @@ directive @withCount(
 GRAPHQL;
     }
 
-    public function batchLoaderClass(): string
+    public function manipulateFieldDefinition(DocumentAST &$documentAST, FieldDefinitionNode &$fieldDefinition, ObjectTypeDefinitionNode &$parentType)
     {
-        return RelationCountBatchLoader::class;
-    }
-
-    public function relationName(): string
-    {
-        $relation = $this->directiveArgValue('relation');
-        if (! $relation) {
-            throw new DefinitionException("You must specify the argument relation in the {$this->name()} directive on {$this->definitionNode->name->value}.");
+        if (RootType::isRootType($parentType->name->value)) {
+            throw new DefinitionException("Can not use @{$this->name()} on fields of a root type.");
         }
 
+        $relation = $this->directiveArgValue('relation');
+        if (! is_string($relation)) {
+            throw new DefinitionException("You must specify the argument relation in the {$this->name()} directive on {$this->definitionNode->name->value}.");
+        }
+    }
+
+    protected function relationName(): string
+    {
+        /**
+         * We validated the argument during schema manipulation.
+         *
+         * @var string $relation
+         */
+        $relation = $this->directiveArgValue('relation');
+
         return $relation;
+    }
+
+    protected function relationLoader(ResolveInfo $resolveInfo): RelationLoader
+    {
+        return new RelationCountLoader(
+            $this->decorateBuilder($resolveInfo)
+        );
     }
 }