+++ src/Schema/Directives/WithDirective.php
@@ -2,10 +2,18 @@
 
 namespace Nuwave\Lighthouse\Schema\Directives;
 
-use Nuwave\Lighthouse\Execution\DataLoader\RelationBatchLoader;
+use GraphQL\Language\AST\FieldDefinitionNode;
+use GraphQL\Language\AST\ObjectTypeDefinitionNode;
+use GraphQL\Type\Definition\ResolveInfo;
+use Nuwave\Lighthouse\Exceptions\DefinitionException;
+use Nuwave\Lighthouse\Execution\DataLoader\RelationLoader;
+use Nuwave\Lighthouse\Execution\DataLoader\SimpleRelationLoader;
+use Nuwave\Lighthouse\Schema\AST\DocumentAST;
+use Nuwave\Lighthouse\Schema\RootType;
+use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 
-class WithDirective extends WithRelationDirective implements FieldMiddleware
+class WithDirective extends WithRelationDirective implements FieldMiddleware, FieldManipulator
 {
     public static function definition(): string
     {
@@ -28,13 +36,23 @@ directive @with(
 GRAPHQL;
     }
 
-    public function batchLoaderClass(): string
+    public function manipulateFieldDefinition(DocumentAST &$documentAST, FieldDefinitionNode &$fieldDefinition, ObjectTypeDefinitionNode &$parentType)
     {
-        return RelationBatchLoader::class;
+        if (RootType::isRootType($parentType->name->value)) {
+            throw new DefinitionException("Can not use @{$this->name()} on fields of a root type.");
+        }
     }
 
-    public function relationName(): string
+    protected function relationName(): string
     {
-        return $this->directiveArgValue('relation', $this->nodeName());
+        return $this->directiveArgValue('relation')
+            ?? $this->nodeName();
+    }
+
+    protected function relationLoader(ResolveInfo $resolveInfo): RelationLoader
+    {
+        return new SimpleRelationLoader(
+            $this->decorateBuilder($resolveInfo)
+        );
     }
 }